Only do clang-format on save when we say we want to
The idea is that you can set the variable 'clang-format-on-save' in
your project's .dir-locals.el file and then your files will be
formatted on save.

This avoids situations where you are editing code in a project that
doesn't have a .clang-format file and you aren't looking to be that
guy that changes every line in every file you touch.